requirements.txt
langchain==0.0.270
psycopg2-binary==2.9.7
langchain_experimental==0.0.10
sqlalchemy==1.4.49
pandas>=1.3.4
openai==0.27.8

pyhive==0.7.0
thrift==0.15.0
thrift-sasl==0.4.3

easydict==1.10
# Flask-Caching==2.0.2
trino==0.326.0
sentence_transformers==2.2.2
InstructorEmbedding==1.0.1
pymongo==4.5.0
redis==5.0.1

text_generation==0.6.1
jupyter==1.0.0
flask==2.3.1
pika==1.3.2
kubernetes==28.1.0
logdecorator==2.4
addict==2.4.0
loguru==0.7.2
#gradio==3.40.0

# vllm==0.1.7
# transformers>=4.33.1
# accelerate==0.21.0
# tiktoken==0.4.0
# huggingface_hub==0.16.4
# bitsandbytes==0.41.1
# sentencepiece==0.1.99
# faiss-gpu==1.7.2
# ray[data,serve]==2.7.0

# 
# # This file will be installed to pod
# #pystan==2.19.1.1
# convertdate==2.3.2
# requests==2.25.1
# click==8.0.0
# numpy>=1.22.2
# pandas>=1.3.4
# LunarCalendar==0.0.9
# #fbprophet>=0.7.1
# scikit-learn==1.2.0
# lightgbm==3.3.0
# catboost==1.0.1
# xgboost==1.6.1
# easydict==1.9
# statsmodels
# psutil==5.8.0
# 
# #feature
# Boruta==0.3
# 
# #nlp
# scipy==1.11.1
# seaborn
# matplotlib
# IPython
# joblib
# ipywidgets
# yellowbrick>=1.0.1
# plotly>=4.4.1
# wordcloud
# textblob
# cufflinks>=0.17.0
# umap-learn
# pyLDAvis
# gensim<=4.3.1
# spacy<2.4.0
# nltk
# mlxtend>=0.17.0
# pyod
# pandas-profiling>=2.8.0
# kmodes>=0.10.1
# imbalanced-learn==0.7.0
# scikit-plot #for lift and gain charts
# pyyaml<=6.0.1
# #numba<0.54
# tqdm
# # streamlit
# numpyencoder
# pmdarima
# keras==2.8.0
# protobuf==3.20.*
# 
# awscli
# mlflow==1.30.0
# importlib_metadata==4.8
# boto3>=1.20.10
# 
# PyHive==0.6.4
# hdfs==2.6.0
# shap
# holidays
# 
# aiohttp
# pyenchant
# simplejson>=3.17.6
# 
# more_itertools
# # jupyter
# 
# ##### Driving Analyzer ##
# shapely
# geopandas
# scikit-image
# #########################
# 
# ##### Time-consuming part package #####
# tensorflow==2.8.0
# pyspark==3.4.0
# kafka-python==2.0.2
# redis==4.4.0
# # eventstoredb==0.7.2
# 
# ##### apt install -yqq gcc g++ libsasl2-modules-gssapi-mit sasl2-bin libsasl2-2 libsasl2-dev libsasl2-modules
# ##### brew install gcc
# sasl>=0.3.1
# pysasl==1.0.0
# thrift==0.15.0
# thrift-sasl==0.4.3
# networkX==2.8.8
# deprecated
# 
# # monitoring
# evidently==0.3.2.1
# 
# umap-learn==0.5.3
